rename(exchange) : ExchangeGetResponseDto 위치 이동
service 패키지에서 dto 패키지로 이동
