+++ src/main/java/com/popcorntalk/domain/exchange/repository/ExchangeRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.popcorntalk.domain.exchange.repository;
 
-import com.popcorntalk.domain.exchange.controller.ExchangeGetResponseDto;
+import com.popcorntalk.domain.exchange.dto.ExchangeGetResponseDto;
 import java.util.List;
 
 public interface ExchangeRepositoryCustom {
